Fixed i386 compilation warnings and errors

+++ library/json_stat.cpp
@@ -25,8 +25,8 @@ struct json_stat_cache {
 
 static void eblob_stat_add_timestamp_raw(rapidjson::Value &stat, const char *name, timeval &tv, rapidjson::Document::AllocatorType &allocator) {
 	rapidjson::Value timestamp(rapidjson::kObjectType);
-	timestamp.AddMember("tv_sec", tv.tv_sec, allocator);
-	timestamp.AddMember("tv_usec", tv.tv_usec, allocator);
+	timestamp.AddMember("tv_sec", (uint64_t)tv.tv_sec, allocator);
+	timestamp.AddMember("tv_usec", (uint64_t)tv.tv_usec, allocator);
 	stat.AddMember(name, allocator, timestamp, allocator);
 }
 
@@ -206,7 +206,7 @@ static void eblob_stat_vfs(struct eblob_backend *b, rapidjson::Value &stat, rapi
 		return;
 	}
 
-	stat.AddMember("bsize", s.f_bsize, allocator);
+	stat.AddMember("bsize", (uint64_t)s.f_bsize, allocator);
 	stat.AddMember("frsize", s.f_frsize, allocator);
 	stat.AddMember("blocks", s.f_blocks, allocator);
 	stat.AddMember("bfree", s.f_bfree, allocator);
@@ -298,7 +298,7 @@ int eblob_json_commit(struct eblob_backend *b) {
 /*
  * calculates lifetime_limit in usecs as doubled periodic timeout
  */
-static uint32_t get_lifetime_limit(struct eblob_backend *b) {
+static long get_lifetime_limit(struct eblob_backend *b) {
 	return b->cfg.periodic_timeout * 2 * 1000000;
 }
